Split off AbstractExtrapolation from collections
